# lib/api-zod/src/generated/types/complaint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dalilak API - Lebanon Accessibility Guide
 * OpenAPI spec version: 0.1.0
 */
import type { ComplaintStatus } from "./complaintStatus";

export interface Complaint {
  id: number;
  placeId?: number;
  placeName?: string;
  message: string;
  senderName: string;
  senderEmail: string;
  status: ComplaintStatus;
  createdAt: string;
}