# lib/api-zod/src/generated/types/complaintBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dalilak API - Lebanon Accessibility Guide
 * OpenAPI spec version: 0.1.0
 */

export interface ComplaintBody {
  placeId?: number;
  message: string;
  senderName: string;
  senderEmail: string;
}