# lib/api-zod/src/generated/types/evaluation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dalilak API - Lebanon Accessibility Guide
 * OpenAPI spec version: 0.1.0
 */

export interface Evaluation {
  id: number;
  placeId: number;
  expertId: number;
  expertName?: string;
  hasRamp: boolean;
  hasElevator: boolean;
  hasAccessibleBathroom: boolean;
  hasAccessibleParking: boolean;
  hasSignLanguage?: boolean;
  hasBraille?: boolean;
  rampCondition?: string;
  notes?: string;
  rating: number;
  createdAt: string;
}