# lib/api-zod/src/generated/types/evaluationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dalilak API - Lebanon Accessibility Guide
 * OpenAPI spec version: 0.1.0
 */
import type { EvaluationBodyRampCondition } from "./evaluationBodyRampCondition";

export interface EvaluationBody {
  hasRamp: boolean;
  hasElevator: boolean;
  hasAccessibleBathroom: boolean;
  hasAccessibleParking: boolean;
  hasSignLanguage?: boolean;
  hasBraille?: boolean;
  rampCondition?: EvaluationBodyRampCondition;
  notes?: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
}