# lib/api-zod/src/generated/types/placeDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dalilak API - Lebanon Accessibility Guide
 * OpenAPI spec version: 0.1.0
 */
import type { Evaluation } from "./evaluation";
import type { Photo } from "./photo";
import type { Place } from "./place";

export type PlaceDetail = Place & {
  photos?: Photo[];
  evaluations?: Evaluation[];
  description?: string;
};