# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dalilak API - Lebanon Accessibility Guide
 * OpenAPI spec version: 0.1.0
 */
import type { Place } from "./place";

export interface Stats {
  totalPlaces: number;
  totalExperts: number;
  totalEvaluations: number;
  totalGovernorates: number;
  recentPlaces: Place[];
}