# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dalilak API - Lebanon Accessibility Guide
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";
import type { UserStatus } from "./userStatus";

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  status: UserStatus;
  phone?: string;
  createdAt: string;
}